feat: reservation controller
