+++ src/main/java/com/dentallink/domain/hospital/dto/request/HospitalUpdateRequest.java
@@ -1,0 +1,4 @@
+package com.dentallink.domain.hospital.dto.request;
+
+public class HospitalUpdateRequest {
+}
