+++ src/main/java/com/dentallink/domain/hospital/dto/response/HospitalUpdateResponse.java
@@ -1,0 +1,4 @@
+package com.dentallink.domain.hospital.dto.response;
+
+public class HospitalUpdateResponse {
+}
