+++ src/main/java/com/dentallink/domain/reservation/controller/ReservationController.java
@@ -1,7 +1,10 @@
 package com.dentallink.domain.reservation.controller;
 
+
 import com.dentallink.common.response.ApiResponse;
 import com.dentallink.common.response.PageResponse;
+import com.dentallink.domain.reservation.dto.AvailableTimeSlotResponse;
+import com.dentallink.domain.reservation.dto.ReservationCreateRequest;
 import com.dentallink.domain.reservation.dto.ReservationResponse;
 import com.dentallink.domain.reservation.dto.ReservationUpdateStatusRequest;
 import com.dentallink.domain.reservation.service.ReservationInternalService;
@@ -12,14 +15,14 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Sort;
 import org.springframework.data.web.PageableDefault;
-import org.springframework.http.HttpStatus;
+import org.springframework.format.annotation.DateTimeFormat;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
-/**
- * TODO: 예약 생성은 HospitalSchedule 완성 후 추가 예정
- */
+import java.time.LocalDate;
+import java.util.List;
+
 @RestController
 @RequestMapping("/api/reservations")
 @RequiredArgsConstructor
@@ -27,6 +30,26 @@ import org.springframework.web.bind.annotation.*;
 public class ReservationController {
 
     private final ReservationInternalService reservationService;
+
+    //예약 생성
+    @PostMapping
+    public ResponseEntity<ApiResponse<ReservationResponse>> createReservation(
+            @RequestBody @Valid ReservationCreateRequest request,
+            @RequestParam @Min(1) Long userId) {
+
+        ReservationResponse response = reservationService.createReservation(request, userId);
+        return ApiResponse.success(response, "예약 성공");
+    }
+
+    //예약 가능 시간 조회
+    @GetMapping("/available-slots")
+    public ResponseEntity<ApiResponse<List<AvailableTimeSlotResponse>>> getAvailableTimePeriod(
+            @RequestParam @Min(1) Long hospitalId,
+            @RequestParam @DateTimeFormat(iso = DateTimeFormat.ISO.DATE) LocalDate date) {
+
+        List<AvailableTimeSlotResponse> availableSlots = reservationService.getAvailableTimePeriod(hospitalId, date);
+        return ApiResponse.success(availableSlots, "예약 가능 시간 조회 성공");
+    }
 
     //예약 단건 조회
     @GetMapping("/{id}")
@@ -48,7 +71,6 @@ public class ReservationController {
         return ApiResponse.pageSuccess(reservations, "내 예약 목록 조회 성공");
     }
 
-
     //병원별 예약 목록 조회
     @GetMapping
     public ResponseEntity<ApiResponse<PageResponse<ReservationResponse>>> getHospitalReservations(
@@ -65,9 +87,7 @@ public class ReservationController {
         return ApiResponse.pageSuccess(reservations, "병원 예약 목록 조회 성공");
     }
 
-    //  예약 생성 (TODO: HospitalSchedule 완성 후 구현)
-
-    //예약 승인
+    //예약 상태 변경 (병원 관리자)
     @PatchMapping("/{id}/status")
     public ResponseEntity<ApiResponse<ReservationResponse>> updateReservationStatus(
             @PathVariable @Min(1) Long id,
@@ -84,7 +104,6 @@ public class ReservationController {
 
     //예약 취소
     @DeleteMapping("/{id}")
-    @ResponseStatus(HttpStatus.NO_CONTENT)
     public ResponseEntity<ApiResponse<Void>> cancelReservation(
             @PathVariable @Min(1) Long id,
             @RequestParam @Min(1) Long userId) {
