+++ src/main/java/com/dentallink/domain/reservation/service/ReservationInternalService.java
@@ -9,10 +9,10 @@ import com.dentallink.domain.reservation.dto.ReservationUpdateStatusRequest;
 import com.dentallink.domain.reservation.entity.Reservation;
 import com.dentallink.domain.reservation.execption.ReservationErrorCode;
 import com.dentallink.domain.reservation.repository.ReservationRepository;
-import com.dentallink.hospital.entity.Hospital;
-import com.dentallink.hospital.entity.HospitalSchedule;
-import com.dentallink.hospital.repository.HospitalRepository;
-import com.dentallink.hospital.repository.HospitalScheduleRepository;
+import com.dentallink.domain.hospital.entity.Hospital;
+import com.dentallink.domain.hospital.entity.HospitalSchedule;
+import com.dentallink.domain.hospital.repository.HospitalRepository;
+import com.dentallink.domain.hospital.repository.HospitalScheduleRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
