+++ src/main/java/com/dentallink/domain/review/controller/ReviewController.java
@@ -1,0 +1,4 @@
+package com.dentallink.domain.review.controller;
+
+public class ReviewController {
+}
