+++ src/main/java/com/dentallink/domain/review/entity/Review.java
@@ -1,0 +1,4 @@
+package com.dentallink.domain.review.entity;
+
+public class Review {
+}
