+++ src/main/java/com/dentallink/domain/review/repository/ReviewRepository.java
@@ -1,0 +1,4 @@
+package com.dentallink.domain.review.repository;
+
+public interface ReviewRepository {
+}
