+++ src/main/java/com/dentallink/domain/review/service/ReviewService.java
@@ -1,0 +1,4 @@
+package com.dentallink.domain.review.service;
+
+public class ReviewService {
+}
