--- src/main/java/com/dentallink/hospital/dto/request/HospitalUpdateRequest.java
@@ -1,4 +1,0 @@
-package com.dentallink.hospital.dto.request;
-
-public class HospitalUpdateRequest {
-}
