--- src/main/java/com/dentallink/hospital/dto/response/HospitalUpdateResponse.java
@@ -1,4 +1,0 @@
-package com.dentallink.hospital.dto.response;
-
-public class HospitalUpdateResponse {
-}
