--- src/main/java/com/dentallink/review/controller/ReviewController.java
@@ -1,4 +1,0 @@
-package com.dentallink.review.controller;
-
-public class ReviewController {
-}
