--- src/main/java/com/dentallink/review/entity/Review.java
@@ -1,4 +1,0 @@
-package com.dentallink.review.entity;
-
-public class Review {
-}
