--- src/main/java/com/dentallink/review/repository/ReviewRepository.java
@@ -1,4 +1,0 @@
-package com.dentallink.review.repository;
-
-public interface ReviewRepository {
-}
