--- src/main/java/com/dentallink/review/service/ReviewService.java
@@ -1,4 +1,0 @@
-package com.dentallink.review.service;
-
-public class ReviewService {
-}
